Move README into documentation (#63)
Leaving the old documentation intact.

Goal: Have a "single source of truth" for eCLM static file generator
documentation.

* collapse deprecated docs using clickable "<details>" tag

* Provide all options for reading the documentation

* Corrected phrasing about the licence

There are different MIT licences.  Usually the Expat licence is used,
but this name is not as commonly used.  Using the undefined article
at least makes the phrasing correct.

* Source environment and set CSMDATA

With moving to the docs/ structure this got lost.  It applies to most
sections, so it must be stated before section 1.

---------

Co-authored-by: Marco van Hulten <Marco.van.Hulten@uni-bonn.de>